20200211
好棒棒

+++ workspace/GitHubTest/src/main/java/com/fund/controller/member/MemberController.java
@@ -30,61 +30,62 @@ public class MemberController {
 	 * @controller 註冊功能
 	 */
 	@RequestMapping(value = "registerMember", method = RequestMethod.POST)
-	public String registerMember(Member member, Model model,@RequestParam(value = "confirmPwd", required = false)String confirmPwd,
+	public String registerMember(Member member, Model model,
+			@RequestParam(value = "confirmPwd", required = false) String confirmPwd,
 			@RequestParam(value = "file", required = false) MultipartFile file, HttpServletRequest request)
 			throws Exception {
-		    model.addAttribute("inputMember", member);//如果輸入錯誤 原本輸入的值導回 使用者不用重複書寫
-			if ((member.getMemberEmail().matches("^[_a-z0-9-]+([.][_a-z0-9-]+)*@[a-z0-9-]+([.][a-z0-9-]+)*$"))
-			 && (member.getMemberPwd().matches( "^(?=^.{6,12}$)((?=.*[0-9])(?=.*[a-z|A-Z]))^.*$"))   && (member.getMemberTwid().matches("^[A-Z]\\d{9}$"))
-			 && (member.getMemberFname() != null) && (member.getMemberSname() != null)
-			 && (member.getMemberGd() != null)    && (member.getMemberHb() != null)
-		     && (member.getMemberCel().matches("^09[0-9]{8}$"))) {
-				if(confirmPwd.equals(member.getMemberPwd())){
-					List<Member> listMember = memberServiceImpl.listFindMemberByEmail(member.getMemberEmail());	
-				if (listMember==null) {//如果帳號不存在				
-					String pic = memberServiceImpl.addMemberPic(file,request);//圖片存檔
-					if(!pic.equals("errorPic")) {//圖片存取正常
-						member.setMemberPic(pic);//設定圖片路徑
-						memberServiceImpl.addMember(member);//存入資料庫
-						return "MemberSystem/save";//註冊成功
-					}else {
-							model.addAttribute("errorPicFormat", "輸入圖片格式錯誤  請確認");
-							return "MemberSystem/register";
-					}} else {
-					List<String> newAccount = memberServiceImpl.suggestAccount(member);//產生建議帳號
+		model.addAttribute("inputMember", member);// 如果輸入錯誤 原本輸入的值導回 使用者不用重複書寫
+		if ((member.getMemberEmail().matches("^[_a-z0-9-]+([.][_a-z0-9-]+)*@[a-z0-9-]+([.][a-z0-9-]+)*$"))
+				&& (member.getMemberPwd().matches("^(?=^.{6,12}$)((?=.*[0-9])(?=.*[a-z|A-Z]))^.*$"))
+				&& (member.getMemberTwid().matches("^[A-Z]\\d{9}$")) && (!member.getMemberFname().isEmpty())
+				&& (!member.getMemberSname().isEmpty()) && (!member.getMemberGd().isEmpty())
+				&& (!member.getMemberHb().isEmpty()) && (member.getMemberCel().matches("^09[0-9]{8}$"))) {
+			if (confirmPwd.equals(member.getMemberPwd())) {
+				List<Member> listMember = memberServiceImpl.listFindMemberByEmail(member.getMemberEmail());
+				if (listMember == null) {// 如果帳號不存在
+					String pic = memberServiceImpl.addMemberPic(file, request);// 圖片存檔
+					if (!pic.equals("errorPic")) {// 圖片存取正常
+						member.setMemberPic(pic);// 設定圖片路徑
+						memberServiceImpl.addMember(member);// 存入資料庫
+						return "MemberSystem/save";// 註冊成功
+					} else {
+						model.addAttribute("errorPicFormat", "輸入圖片格式錯誤  請確認");
+						return "MemberSystem/register";
+					}
+				} else {
+					List<String> newAccount = memberServiceImpl.suggestAccount(member);// 產生建議帳號
 					model.addAttribute("errorAccount", newAccount);
-					System.out.println("帳號已註冊   以下提供參考"+newAccount);
+					System.out.println("帳號已註冊   以下提供參考" + newAccount);
 					return "MemberSystem/register";// 帳號已註冊
 				}
-				}else {
-					System.out.println("兩次密碼輸入不符合");
-					model.addAttribute("errorPwd", "兩次密碼輸入不符合" );
-					return "MemberSystem/register";
-				}		
 			} else {
-				System.out.println("資料輸入格式錯誤");
-				model.addAttribute("errorFormat", "資料輸入格式錯誤" );
+				System.out.println("兩次密碼輸入不符合");
+				model.addAttribute("errorPwd", "兩次密碼輸入不符合");
 				return "MemberSystem/register";
 			}
-		} 
-	
+		} else {
+			System.out.println("資料輸入格式錯誤");
+			model.addAttribute("errorFormat", "資料輸入格式錯誤");
+			return "MemberSystem/register";
+		}
+	}
 
 	/**
 	 * @controller 登入功能
 	 */
 	@RequestMapping(value = "/loginMember", method = RequestMethod.POST)
-	public String loginMember(Member member, Model model ,HttpSession httpSession) {
+	public String loginMember(Member member, Model model, HttpSession httpSession) {
 		String memberEmail = member.getMemberEmail();
-		String pwd = DigestUtils.md5DigestAsHex(member.getMemberPwd().getBytes());//md5解密
-		if (memberEmail.matches("^[_a-z0-9-]+([.][_a-z0-9-]+)*@[a-z0-9-]+([.][a-z0-9-]+)*$") && pwd != null) {
+		String pwd = DigestUtils.md5DigestAsHex(member.getMemberPwd().getBytes());// md5解密
+		if (memberEmail.matches("^[_a-z0-9-]+([.][_a-z0-9-]+)*@[a-z0-9-]+([.][a-z0-9-]+)*$") && !pwd.isEmpty()) {
 			List<Member> listMember = memberServiceImpl.listFindMemberByEmail(memberEmail);
-			if (listMember!=null) {
+			if (listMember != null) {
 				String rpwd = listMember.get(0).getMemberPwd();
 				if (pwd.equals(rpwd)) {
 					int memberId = listMember.get(0).getMemberId();
 					Member memberInformation = memberServiceImpl.findMember(memberId);
-					httpSession.setAttribute("memberInformation", memberInformation);//memberInformation 存入Session
-					memberServiceImpl.addStoreIdSession(memberId, httpSession);//StoreId 存入Session
+					httpSession.setAttribute("memberInformation", memberInformation);// memberInformation 存入Session
+					memberServiceImpl.addStoreIdSession(memberId, httpSession);// StoreId 存入Session
 					System.out.println("登入成功 ");
 					return "home";
 				} else {
@@ -111,8 +112,8 @@ public class MemberController {
 	public String getMember(HttpSession httpSession, Model model) {
 		Member memberSession = (Member) httpSession.getAttribute("memberInformation");// 取得Session的Member
 		if (memberSession != null) {
-			Member member = memberServiceImpl.findMember(memberSession.getMemberId());//查資料庫會員資料
-			model.addAttribute("getMember",member);
+			Member member = memberServiceImpl.findMember(memberSession.getMemberId());// 查資料庫會員資料
+			model.addAttribute("getMember", member);
 			return "MemberSystem/updateMember";
 		} else {
 			model.addAttribute("nologin", "請先登入 3秒後跳至首頁");
@@ -127,33 +128,34 @@ public class MemberController {
 	public String memberUpdate(HttpSession httpSession, Member member, Model model,
 			@RequestParam(value = "file", required = false) MultipartFile file, HttpServletRequest request)
 			throws Exception {
-		if ((member.getMemberTwid().matches("^[A-Z]\\d{9}$"))&& (!member.getMemberFname().isEmpty()) 
-			 && (!member.getMemberSname().isEmpty())&& (!member.getMemberGd().isEmpty()) 
-			 && (!member.getMemberHb().isEmpty()) && (member.getMemberCel().matches("^09[0-9]{8}$"))) {
+		if ((member.getMemberTwid().matches("^[A-Z]\\d{9}$")) && (!member.getMemberFname().isEmpty())
+				&& (!member.getMemberSname().isEmpty()) && (!member.getMemberGd().isEmpty())
+				&& (!member.getMemberHb().isEmpty()) && (member.getMemberCel().matches("^09[0-9]{8}$"))) {
 			Member memberSession = (Member) httpSession.getAttribute("memberInformation");// 取得Session的Member
-		if (memberSession != null) {
-			if (!file.getOriginalFilename().isEmpty()) {//檢驗看圖檔是否有上傳 
-				String fileType = file.getContentType(); // 獲得檔案型別
-				if (fileType.equals("image/jpeg") || fileType.equals("image/gif")) {//查看圖片資料類型
-				memberServiceImpl.deleteMemberPic(memberSession.getMemberId(), request);//刪除原本照片
-				String path = memberServiceImpl.addMemberPic(file, request);//新增照片
-				member.setMemberPic(path);// 把圖片儲存路徑儲存到資料庫
-				memberServiceImpl.updateMember(member, memberSession.getMemberId());//進行更新
-				}else {
-					System.out.println("輸入格式錯誤");
-					return "MemberSystem/error";
+			if (memberSession != null) {
+				if (!file.getOriginalFilename().isEmpty()) {// 檢驗看圖檔是否有上傳
+					String fileType = file.getContentType(); // 獲得檔案型別
+					if (fileType.equals("image/jpeg") || fileType.equals("image/gif")) {// 查看圖片資料類型
+						memberServiceImpl.deleteMemberPic(memberSession.getMemberId(), request);// 刪除原本照片
+						String path = memberServiceImpl.addMemberPic(file, request);// 新增照片
+						member.setMemberPic(path);// 把圖片儲存路徑儲存到資料庫
+						memberServiceImpl.updateMember(member, memberSession.getMemberId());// 進行更新
+					} else {
+						System.out.println("輸入格式錯誤");
+						return "MemberSystem/error";
+					}
+				} else {
+					memberServiceImpl.updateMember(member, memberSession.getMemberId());// 沒修改照片話直接更新
 				}
+				return "redirect:/getMember";// 更新完成後 返回getMember 方法 導回會員查詢頁面
 			} else {
-				memberServiceImpl.updateMember(member, memberSession.getMemberId());//沒修改照片話直接更新
+				return "MemberSystem/noLogin";
 			}
-			return "redirect:/getMember";//更新完成後 返回getMember 方法  導回會員查詢頁面
 		} else {
-			return "MemberSystem/noLogin";
-		}}else {
 			System.out.println("輸入格式錯誤");
 			return "MemberSystem/error";
 		}
-		
+
 	}
 
 	/**
@@ -164,8 +166,8 @@ public class MemberController {
 		Member memberSession = (Member) httpSession.getAttribute("memberInformation");// 取得Session的Member
 		if (memberSession != null) {
 			Member member = memberServiceImpl.findMember(memberSession.getMemberId());
-			httpSession.setAttribute("memberInformation", member);//將更改後memberInformation Session放入
-			memberServiceImpl.addStoreIdSession(member.getMemberId(), httpSession);//StoreId 存入Session
+			httpSession.setAttribute("memberInformation", member);// 將更改後memberInformation Session放入
+			memberServiceImpl.addStoreIdSession(member.getMemberId(), httpSession);// StoreId 存入Session
 			model.addAttribute("getMember", member);
 			return "MemberSystem/mbInformation";
 		} else {
@@ -173,29 +175,7 @@ public class MemberController {
 			return "MemberSystem/noLogin";
 		}
 	}
-	
-	public String findPwd() {
-		
-		return "";
-	}
-	
 
-	
-	/**
-	 * @映射路徑到register
-	 */
-	@RequestMapping(path="/register")
-	public String register() {
-		return "MemberSystem/register";
-	}
-	
-	/**
-	 * @映射路徑到loginSystem
-	 */
-	@RequestMapping(path="/loginSystem")
-	public String loginSystem() {
-		return "MemberSystem/loginSystem";
-	}
 
 
 }
